refactor: n+1
