+++ api-under-test/Controllers/Challenge2Controller.cs
@@ -14,7 +14,8 @@ namespace api_under_test.Controllers
     [ApiController]
     [Route("weatherforecast_challenge2")]
     public class Challenge2Controller: ControllerBase
-    { private static readonly string[] Summaries = new[]
+    {
+        private static readonly string[] Summaries = new[]
         {
             "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
         };
@@ -31,31 +32,26 @@ namespace api_under_test.Controllers
         {
            var chaosPolicy = MonkeyPolicy.InjectLatencyAsync(with =>
                 with.Latency(TimeSpan.FromSeconds(1))
-                    .InjectionRate(0.05) // 10 % 
+                    .InjectionRate(0.1) // 10 % 
                     .Enabled(true));    // Would probably only turn it on in some environments
             var mix = Policy.WrapAsync(GetPolicy(), chaosPolicy);
             return await mix.ExecuteAsync((ct) => GetForecasts(ct), CancellationToken.None);
         }
 
         private IAsyncPolicy GetPolicy() {
-            // you can change from here 
-            var retries = 0;
-            var timeout = TimeSpan.FromMilliseconds(3000);
+            // Change these:
+            var timeout = TimeSpan.FromMilliseconds(30000);
             Program.ConfiguredTimeout.Set(timeout.TotalMilliseconds);
-            Program.ConfiguredRetries.Set(retries);
 
             Program.ConfiguredRetries.Publish();
             Program.ConfiguredTimeout.Publish();
 
-            var retryPolicy = Policy.Handle<Exception>().RetryAsync(retries, (ex, attempt) => Program.ExecutedRetries.Inc());
-            var timeoutPolicy = Policy.TimeoutAsync(timeout);
-            return Policy.WrapAsync(timeoutPolicy, retryPolicy);
-            // until here
+            return Policy.TimeoutAsync(timeout);
         }
 
         private async Task<IEnumerable<WeatherForecast>> GetForecasts(CancellationToken ct)
         {
-            await Task.Delay(1, ct);
+            await Task.Delay(20, ct);
             var rng = new Random();
             return Enumerable.Range(1, 5).Select(index => new WeatherForecast
             {
