+++ api-under-test/Controllers/Challenge4Controller.cs
@@ -5,9 +5,10 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Polly;
-using System.Threading;
 using Polly.Contrib.Simmy;
 using Polly.Contrib.Simmy.Latency;
+using Polly.Contrib.Simmy.Outcomes;
+using System.Threading;
 
 namespace api_under_test.Controllers
 {
@@ -21,62 +22,59 @@ namespace api_under_test.Controllers
         };
 
         private readonly ILogger<Challenge4Controller> _logger;
-        private readonly Random _rng; 
 
         public Challenge4Controller(ILogger<Challenge4Controller> logger)
         {
-            _rng = new Random();
             _logger = logger;
         }
 
         [HttpGet]
         public async Task<IEnumerable<WeatherForecast>> Get()
         {
-            var policy = GetPolicy();
-            // Unfortunately, in the real world there is a little more to cancellation.
-            // It is explained here: 
-            // https://github.com/App-vNext/Polly/wiki/Timeout
-            // The example with CancellationToken.None shows a way to implement it
-            // if you want to provide a new cancellation token that executeasync can use to terminate a task
-            // The thing is, tasks must collaborate on cancellation
-            // and they do that by sending cancellationtokens... They must come from somewhere 
-            return await policy.ExecuteAsync(() => GetForecasts());
+           Exception fault = new System.Net.Sockets.SocketException(errorCode: 10013);
+           var latencyMonkey = MonkeyPolicy.InjectLatencyAsync(with =>
+                with.Latency(TimeSpan.FromSeconds(1))
+                    .InjectionRate(0.1) 
+                    .Enabled(true));
+
+            var errorMonkey = MonkeyPolicy.InjectExceptionAsync(with => 
+                with.Fault(fault)
+                    .InjectionRate(0.1)
+                    .Enabled(true));
+ 
+            var monkeyPolicy = Policy.WrapAsync(latencyMonkey, errorMonkey);
+
+            var mix = Policy.WrapAsync(GetPolicy(), monkeyPolicy);
+            return await mix.ExecuteAsync((ct) => GetForecasts(ct), CancellationToken.None);
         }
 
         private IAsyncPolicy GetPolicy() {
-            var timeoutPolicy = Policy.TimeoutAsync(TimeSpan.FromMilliseconds(20));
-            return timeoutPolicy; 
+            // you can change from here 
+            var retries = 0;
+            var timeout = TimeSpan.FromMilliseconds(3000);
+            Program.ConfiguredTimeout.Set(timeout.TotalMilliseconds);
+            Program.ConfiguredRetries.Set(retries);
+
+            Program.ConfiguredRetries.Publish();
+            Program.ConfiguredTimeout.Publish();
+
+            var retryPolicy = Policy.Handle<Exception>().RetryAsync(retries, (ex, attempt) => Program.ExecutedRetries.Inc());
+            var timeoutPolicy = Policy.TimeoutAsync(timeout);
+            return Policy.WrapAsync(retryPolicy, timeoutPolicy);
+            // until here
         }
 
-        // This signature needs to change to accept a CancellationToken 
-        private async Task<IEnumerable<WeatherForecast>> GetForecasts()
+        private async Task<IEnumerable<WeatherForecast>> GetForecasts(CancellationToken ct)
         {
-            // And this too must propagate tokens 
-            var tasks = Enumerable.Range(1, 3).Select(index => GetForecast(index)).ToArray();
-            await Task.WhenAll(tasks);
-            return tasks.Select(t => t.Result).ToArray();
-        }
-
-        // The CancellationToken needs to go here as well 
-        private async Task<WeatherForecast> GetForecast(int index){
-            
-            // The delays need to cooperate with the cancellation by being handled a token 
-            // https://docs.microsoft.com/en-us/dotnet/api/system.threading.tasks.task.delay?view=netcore-3.1
-
-            // (Also other async methods in .NET accept CancellationTokens, for example:
-            // https://docs.microsoft.com/en-us/dotnet/api/system.net.http.httpclient.sendasync?view=netcore-3.1)
-            if (_rng.NextDouble() > 0.95) {
-               await Task.Delay(2000);
-            } else {
-                await Task.Delay(10);
-            }
-
-            return new WeatherForecast
+            await Task.Delay(20, ct);
+            var rng = new Random();
+            return Enumerable.Range(1, 5).Select(index => new WeatherForecast
             {
                 Date = DateTime.Now.AddDays(index),
-                TemperatureC = _rng.Next(-20, 55),
-                Summary = Summaries[_rng.Next(Summaries.Length)]
-            };
-       }
+                TemperatureC = rng.Next(-20, 55),
+                Summary = Summaries[rng.Next(Summaries.Length)]
+            })
+            .ToArray();
+        }
     }
 }
