+++ api-under-test/Controllers/Challenge5Controller.cs
@@ -30,11 +30,17 @@ namespace api_under_test.Controllers
         }
 
         [HttpGet]
-        public async Task<IEnumerable<WeatherForecast>> Get(CancellationToken ct)
+        public async Task<IEnumerable<WeatherForecast>> Get()
         {
             var policy = GetPolicy();
-            // We can also use someone elses token, the rest should be quite similar to challenge 4
-            return await policy.ExecuteAsync(() => GetForecasts()); 
+            // Unfortunately, in the real world there is a little more to cancellation.
+            // It is explained here: 
+            // https://github.com/App-vNext/Polly/wiki/Timeout
+            // The example with CancellationToken.None shows a way to implement it
+            // if you want to provide a new cancellation token that executeasync can use to terminate a task
+            // The thing is, tasks must collaborate on cancellation
+            // and they do that by sending cancellationtokens... They must come from somewhere 
+            return await policy.ExecuteAsync(() => GetForecasts());
         }
 
         private IAsyncPolicy GetPolicy() {
@@ -42,7 +48,7 @@ namespace api_under_test.Controllers
             return timeoutPolicy; 
         }
 
-        // This signature needs to change to accept a token 
+        // This signature needs to change to accept a CancellationToken 
         private async Task<IEnumerable<WeatherForecast>> GetForecasts()
         {
             // And this too must propagate tokens 
@@ -51,11 +57,14 @@ namespace api_under_test.Controllers
             return tasks.Select(t => t.Result).ToArray();
         }
 
-        // There might be some changes required to this signature as well
+        // The CancellationToken needs to go here as well 
         private async Task<WeatherForecast> GetForecast(int index){
             
             // The delays need to cooperate with the cancellation by being handled a token 
+            // https://docs.microsoft.com/en-us/dotnet/api/system.threading.tasks.task.delay?view=netcore-3.1
 
+            // (Also other async methods in .NET accept CancellationTokens, for example:
+            // https://docs.microsoft.com/en-us/dotnet/api/system.net.http.httpclient.sendasync?view=netcore-3.1)
             if (_rng.NextDouble() > 0.95) {
                await Task.Delay(2000);
             } else {
