+++ api-under-test/Controllers/Challenge6Controller.cs
@@ -8,8 +8,6 @@ using Polly;
 using System.Threading;
 using Polly.Contrib.Simmy;
 using Polly.Contrib.Simmy.Latency;
-using Polly.Contrib.Simmy.Outcomes;
-using System.Net.Http;
 
 namespace api_under_test.Controllers
 {
@@ -24,38 +22,52 @@ namespace api_under_test.Controllers
 
         private readonly ILogger<Challenge6Controller> _logger;
         private readonly Random _rng; 
-        private readonly HttpClient _client; 
 
         public Challenge6Controller(ILogger<Challenge6Controller> logger)
         {
             _rng = new Random();
             _logger = logger;
-            _client = new HttpClient();
         }
 
         [HttpGet]
-        public async Task<string> Get()
+        public async Task<IEnumerable<WeatherForecast>> Get(CancellationToken ct)
         {
             var policy = GetPolicy();
-            // these lines might need to be moved around
-            var url = new Uri(@"https://localhost:5001/weatherforecast_challenge0");
-            var msg = new HttpRequestMessage(HttpMethod.Get, url); 
-            // to somewhere... and the GetForecasts signature got to loose its message
-            return await policy.ExecuteAsync(() => GetForecasts(msg)); 
+            // We can also use someone elses token, the rest should be quite similar to challenge 5
+            return await policy.ExecuteAsync(() => GetForecasts()); 
         }
 
         private IAsyncPolicy GetPolicy() {
-            return Policy.Handle<Exception>().RetryAsync(3);
+            var timeoutPolicy = Policy.TimeoutAsync(TimeSpan.FromMilliseconds(20));
+            return timeoutPolicy; 
         }
 
-        private async Task<string> GetForecasts(HttpRequestMessage msg)
+        // This signature needs to change to accept a token 
+        private async Task<IEnumerable<WeatherForecast>> GetForecasts()
         {
-               var resp = await _client.SendAsync(msg);
-               if (_rng.NextDouble() > 0.95) {
-                    throw new Exception( "oops");
-               }
+            // And this too must propagate tokens 
+            var tasks = Enumerable.Range(1, 3).Select(index => GetForecast(index)).ToArray();
+            await Task.WhenAll(tasks);
+            return tasks.Select(t => t.Result).ToArray();
+        }
 
-               return await resp.Content.ReadAsStringAsync(); 
-        } 
-    } 
-} 
+        // There might be some changes required to this signature as well
+        private async Task<WeatherForecast> GetForecast(int index){
+            
+            // The delays need to cooperate with the cancellation by being handled a token 
+
+            if (_rng.NextDouble() > 0.95) {
+               await Task.Delay(2000);
+            } else {
+                await Task.Delay(10);
+            }
+
+            return new WeatherForecast
+            {
+                Date = DateTime.Now.AddDays(index),
+                TemperatureC = _rng.Next(-20, 55),
+                Summary = Summaries[_rng.Next(Summaries.Length)]
+            };
+       }
+    }
+}
