+++ k6-tests/challenge2test.js
@@ -18,7 +18,7 @@ const myOkRate = new Rate("200 OK rate");
 const myOkCounter = new Counter("200 OK count");
 
 export default function() {
-  let response = http.get("http://localhost:5000/weatherforecast_challenge2");
+  let response = http.get("http://localhost:5000/weatherforecast_challenge3");
   let resOk = response.status === 200;
   myOkRate.add(resOk);
   myOkCounter.add(resOk);
