+++ k6-tests/challenge3test.js
@@ -19,7 +19,7 @@ const myOkRate = new Rate("200 OK rate");
 const myOkCounter = new Counter("200 OK count");
 
 export default function() {
-  let response = http.get("http://localhost:5000/weatherforecast_challenge3");
+  let response = http.get("http://localhost:5000/weatherforecast_challenge4");
   let resOk = response.status === 200;
   TrendRTT.add(response.timings.duration);
   myOkRate.add(resOk);
