+++ k6-tests/challenge4test.js
@@ -24,7 +24,7 @@ const myOkRate = new Rate("200 OK rate");
 const myOkCounter = new Counter("200 OK count");
 
 export default function() {
-  let response = http.get("http://localhost:5000/weatherforecast_challenge4");
+  let response = http.get("http://localhost:5000/weatherforecast_challenge5");
   let resOk = response.status === 200;
   myOkRate.add(resOk);
   myOkCounter.add(resOk);
