+++ k6-tests/challenge6test.js
@@ -19,7 +19,7 @@ const myOkRate = new Rate("200 OK rate");
 const myOkCounter = new Counter("200 OK count");
 
 export default function() {
-  let response = http.get("http://localhost:5000/weatherforecast_challenge6");
+  let response = http.get("http://localhost:5000/weatherforecast_challenge7");
   let resOk = response.status === 200;
   myOkRate.add(resOk);
   myOkCounter.add(resOk);
